PDO based queues: Fix excessively low poll interval, and make it configurable (fix for strict PHP type-checks)

+++ src/Pdo/Queue.php
@@ -125,7 +125,7 @@ class Queue implements QueueInterface
                 continue;
             }
 
-            usleep($pollIntervalSecs * 1000000);
+            usleep((int) ($pollIntervalSecs * 1000000));
         }
     }
